Add task ID to expanded list item
This ID is useful to locate the task directory for inspection.

Because it is fairly long, it seemed better to keep it relatively
low in the list of properties to avoid making the essential information
more difficult to read or skim through.

Yet placing it above the task output minimizes movement when the console
is toggled open.

See https://community.opendronemap.org/t/19285